requirements.txt
requests>=2.25.0
aiohttp>=3.8.0
asyncio>=3.4.3
matplotlib>=3.4.0
pandas>=1.2.0
PyYAML>=5.4.0
python-dateutil>=2.8.1
